flask/requirements.txt
certifi==2019.6.16
Click==7.0
cycler==0.10.0
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
nltk==3.4.5
numpy==1.17.2
Pillow==6.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
six==1.12.0
torch==1.2.0
torchvision==0.4.0
Werkzeug==0.15.6
